Migrate article edit events from E3Experiments
Rather than issue a separate event for article creation,
decided to simply set a flag. Turns out that checking for
the parent rev ID being null is adequate for detecting
article creation. I even tested it by deleting an article
and recreating it.

I would have also liked to get rid of the API hook handler,
if only there was some easy way of detecting API calls in
ArticleSaveComplete.

If you want to verify this, you'll need to declare
$wgEventLoggingFile in LocalSettings.php. Like the old
ClickTracking log, it can be a filename.

Change-Id: I87cc40daa7c8de07a8993911a3d469c5f0a4030f