Clean up: Document baseUri, fix test failure.
Base url is now properly documented, in particular as no longer
being assumed to have a trailing '?'.

Previous code assumed the string had a trailing '?' which was
rather unintuitive. Backwards compatibility kept through onSetup
hook (to make deployment easier).

test.setDefaults has an asynchronous aspect, the test was
previously failing (except when the base url was set to
null, in which case no request was made and the deferred
resolved synchronously).

Change-Id: I13f3490ae279e8e85464ad207b5ddbe036e5d95a