Raw varnish log -> schema-validated JSON stream
This contains a whole pile of ugly hacks I'm not especially proud of. A
lot of them were implemented to 'fix' the loss of type information when
encoding JSON to query string. Once change 40554 is merged and deployed,
we will be able to send raw JSON as a single percent-encoded query param
and thereby not lose type information nor be constrained to compress
field names. I will then go back and simplify this code.

Change-Id: Ice4127040c360b234a0abd685d0cc74fd57ffe02