Remove unnecessary type-coercion from account creation event
I wavered about whether to keep the type-coercion or not. I previously
reasoned that the type-coercion doubles as a kind of annotation that
makes the code easier to reason about, but I think it does the opposite,
since it signals defensiveness and lack of confidence about the
values we're handling.

Change-Id: I31b6cca8a942680108896d1954472ab6908e73e3