Re-create table if it was altered or dropped
Previously I had assumed that if the table exists in SQLAlchemy's Meta,
it must also exist in the database. This does not take into account the
possibility that someone other than json2sql goes and drops or alters
the table behind its back. I would chalk this up as pathological
behavior that should not be caught, but I think it may actually be a
viable strategy for archiving old data. (i.e., rename table foo to
foo_old and rely on json2sql to recreate the table for new events).

Change-Id: Ib25eacd5b0305029253c8ef2731f8d8f8f2d185c