Prevent defaults from clobbering event properties
As Matt Flaschen noted in 48248, the event object and the defaults object are
passed to $.extend() in the wrong order, causing the defaults to clobber event
data for properties that are defined in both objects. This change reversed the
order of the two parameters, causing event properties to clobber the defaults,
as intended. The change also updates one of the test cases to cover this
particular scenario.

Also corrects comment which advertises a feature that was not implemented (or
was implemented and then reverted -- can't remember).

Bug: 48248
Change-Id: Id14ba276329272c41370458d5330599a725f5419