Standardize executable names; bump version to 0.6
In anticipation of a re-write of the EventLogging Puppet module, I am changing
the names of the executable scripts so that they follow a clear convention
(namely, starting with/'eventlogging' prefix and using full, hyphen-separated
words).

This change also deletes some obsoleted scripts. 'log2json', 'json2mongo', and
'json2sql' are deprecated in favor of the generic 'eventlogging-consumer',
which can write to different back-ends. 'sv-alerts' is dependent on supervisor,
which we're migrating away from; it needs to be replaced w/robust icinga
support.

Change-Id: Iada1a398e3044c44260c7f3eee74cc546e7ad08c