Use params->kwargs mapping to determine database name
"map['database']" is a remnant of code that mostly removed in
change I9c12ee6f2. Before that change, 'map' was a mapping of arguments
provided by pymongo.uri_parser. After it, map was not getting assigned, and
thus defaulted to the built-in map method.

The correct way to receive arguments specified via the URI is to rely on
deparam & call to map the URI params to kwargs.

Change-Id: I7882b2a9877aac50069ec2349a236b39abe9f42a