Have mtime as calculated by startup module increase on schema change
ResourceLoaderStartupModule calculates the effective mtime as the
maximum of the module's mtime, and $wgCacheEpoch (in UNIX time).
The latter was always greater, so schema updates did not increase
the effective revid.

This fixes that by ensuring the schema module always returns an mtime
greater than $wgCacheEpoch.

There is a new test for this (and updates to an old one), as well as
some related refactoring of the test class.

Bug: 60942
Change-Id: I3de7c406f86838943d23a44da36b8790b2b2bbad