Allow schema registration via custom hook
This patch adds a custom hook, 'EventLoggingRegisterSchemas', that extensions
can use to declare schemas. Hook handlers are called with an associative array
parameter, passed by reference, that maps schema names to revision IDs. A
simple handler for this hook may look like this:

  $wgHooks[ 'EventLoggingRegisterSchemas' ][] = function ( &$schemas ) {
    $schemas[ 'MultimediaViewerNetworkPerformance' ] = 7917896;
  };

The advantage of this approach that it does not introduce a hard dependency on
the EventLogging extension. If the extension is not present, the hook never
fires.

Schemas registered in this fashion can obviously not be declared as
dependencies for ResourceLoader modules that are declared unconditionally.
Instead, the intention is that developers migrate to using mw#track to log
events, relying on the implementation provided in change I0cee7a03e.

Change-Id: I2d29315f9a4d8fa2135c97584469c2842c1071be