Add $wgEventLoggingSchemas
In simple cases, schema registration does not even need to happen in a hook
handler. This patch adds a configuration variable, $wgEventLoggingSchemas,
which is an associative array mapping schema names to revision IDs. Extensions
can add members to this array, which will cause them to be registered as schema
modules by EventLogging if the EventLogging extension is enabled. Since I
expect this method of registering schemas to represent the common case, I
updated the generated code sample on schema pages to reflect it.

Change-Id: Icb72ba325f1d10bff748c13df8e5276f56281b8e