Do not override existing value when initializing $wgEventLoggingSchemas
This makes it possible for extensions to set $wgEventLoggingSchemas
before it is being initialized by the EventLogging extension.

This might or might not be a good idea. In PHP,

  $wgEventLoggingSchemas['foo'] = 123;

is entirely valid even if $wgEventLoggingSchemas has not been defined
yet, so this patch would make it possible to set that variable
in a convenient way, just like other globals are set, without
resorting to hooks or callbacks. On the other hand,

  $wgEventLoggingSchemas += array(
      'foo' => 123,
      'bar' => 456,
  );

would still break, so this might give developers a false sense of
security.

Bug: 65680
Change-Id: I3b4680f162ad3b1796672dae5a50a230beae0fdc