Add a monotonic clock implementation
I needed a monotonic clock for a change I wanted to make to PeriodicThread.
Python 3.3 introduced `time.monotonic`, but there's nothing equivalent for
earlier versions of Python. So I set out to write a fallback implementation.
One thing led to another, and I ended up with a small, self-standing library,
which I published to GitHub and PyPI:

- https://github.com/atdt/monotonic
- https://pypi.python.org/pypi/monotonic

So now there's the question of how to integrate it with EventLogging. We could
treat it like a normal external dependency, but that's a lot of inconvenience
for a library that fits in a single file and of which I am the primary author.
So I decided to just bundle it here instead.

Change-Id: If6166d2353f51d30a447b0a6405dbe8871431502