Make sql_writer accept a 'replace' parameter
By adding the parameter to sql_writer's signature, we ensure that it will be
passed through by eventlogging.factory.apply_safe.

Because SQLAlchemy decodes URL parameters to kwargs for the DBAPI's connect()
method, we have to strip out the parameter from the URI before passing it to
sqlalchemy.create_engine().

Usage example:

  eventlogging-consumer tcp://localhost:8600 mysql://user:pass@db?replace=1

Change-Id: I89776b4af4e82d2e2f806ffb60bf8489306e50d7