Harmonize logging setup
Each of the scripts set up logging on their own. And it seems the
setups differed not due to some special need, but because it's hard to
keep them in sync. Also, it was especially cumbersome that
eventlogging-consumer does not add a timestamp. This made
reconstructing issues from only the logs unnecessarily hard. Hence, we
now setup up the scripts' logging in a central place. That's easier to
maintain, and buys us timestamps in eventlogging-consumer instances
for free.

Change-Id: I600a6dfbcdf65c70789f35e59b7b65039625aeda