Add a simple client API
My motivation with this change is to make EventLogging be a useful library
(rather than just a standalone app) for authors of scripts that process events.
To that end, this change adds an EventConsumer class (aliased as
'eventlogging.connect') that makes it easy to consume events from Python code.

Example usage:

  import eventlogging

  events = eventlogging.connect('tcp://vanadium.eqiad.wmnet:8600')
  for event in events.filter(schema='NavigationTiming'):
    # ...

Change-Id: I3ce7837b15855d077424382c15fbf765441f5f92