Remove clientValidated, isTruncated from event
We have deployed changes to remove clientValidated and isTruncated
from the events sent by the javascript client. Because these changes
do not get deployed to all projects at the same time
we actively remove those fields if they are present on the event.
That way the event validates against the present capsule that does no
longer include both fields.

Need to test in beta labs

Change-Id: I74fa9a8e67df710febf98b687f5dc39195263325