Add size limit to event url
Today, when the event url is too long, the query string gets
truncated to its first 1014 chars by varnishncsa. This results
in validation errors that can be frequent if the schema is
sufficiently big. The consequence is unnecessary overhead in
the EventLogging back-end and unpredictible variations in the
validation alerting system.

This change adds a check on the client-side to avoid sending
events that are too long. In that case an error is logged in
the console if available. It also adds a check in the dev-server.

Bug: T91918
Change-Id: Ie164a9357cdd599823f64122dcc79fec2198832f