Statsd reporter checks format of processor files
We added a new processor of files whose format doesn't exactly abide
to what the statsd reporter is expecting. The new processor format
needs to be added. Now, since reporting of metrics
and thus, alarms are broken in prod this change just fixes the reporter so it
continues to report the pre-existing metrics.

Change-Id: I515d1362e5c3183d61e4b8cf7f129e2f5fca3348