tests: Remove obsolete cache-epoch assertion
The assertion isn't useful anymore since 63c1c83. That commit
removed the cache-epoch logic from the SchemaModule class.

The test was still passing because the hash timestamp is
generated fresh on each test run (and therefore higher than
whatever epoch is configured to in MediaWiki), but wasn't doing
anything useful.

Removing this test because it would otherwise be failing after
change Ibb292d2416839327d lands in MediaWiki. Because the
getDefinitionMtime() helper will start returning dummy values and
shouldn't be relied upon in this fashion either way.

Change-Id: I68bfd0d9e52ac7277862c982ddc1b99894548e7f