Update default api uri to use https
I was having issues getting schema's to load in my vagrant
instance, it turns out this http link is redirecting to
https but the http client used by EventLogging doesn't follow
these redirects.

Could alternatively add some option for the http client to
follow redirects, but since the canonical uri is https that
can be delt with separately from this patch.

Change-Id: Ic344820ff9ffd2ec5c114b9e20557d16a5f84805