Provide inSample() method
Currently this is hardcoded in all WikimediaEvents consumers and elsewhere.

This is implemented as part of the subscriber module instead of the core module
so that consumers can use this in the boostrapping code (which is always loaded)
instead of loading the core module first.

Based on implementation from I0cb6fe37e.

Tested in Chrome, Safari and Firefox. Both with Crypto API and with Math.random()
in mediawiki.user.

Also phase out the ext.eventLogging.init and ext.eventLogging.Schema modules
introduced in 306f4c1.

Bug: T112603
Change-Id: I4a4f06ae195fb331762eadacd16c42322ea22882