Add option to see EventLogging happen on page views
This simplifies some debugging work, along with showing how an
eventlogging schema works to other team members such as analysts
that will be reviewing the data.

It also makes it possible for normal users to see how they are
being tracked, although I'm not sure the right place to expose
it in the user interface. As implemented this is mostly a debug
flag because it needs to be manually turned on for a user rather
than enabled from the web interface.

To enable, run the following in the console:

  mw.loader.using('mediawiki.api.options').then(function () {
    new mw.Api().saveOption('eventlogging-display-web', '1');
  });

With WikimediaEvents installed, an easy way to trigger it is:

  mw.track('mw.deprecate', 'foo');

Change-Id: I1ac4a55762367b8433d203e2dce4d2108aff3797