ext.eventLogging.subscriber: Don't wait for "load" if it already happened
On wikis with jQuery 3 enabled, I noticed it sometimes prints this warning:

> JQMIGRATE: jQuery(window).on('load'...) called after load event occurred

Given "load" is an event, and given browsers don't fire event handlers
retroactively, this means on those page views, we end up never subscribing to
mw.track events.

https://github.com/jquery/jquery-migrate/blob/c12e1ab6/warnings.md

This isn't a regression in jQuery 3, but they do add a nice warning
as of this version to detect this common mistake.

Note that $()/$(document).ready() isn't affected because those handlers
are Promise-based (which always trigger callbacks retroactively if said
Promise is already resolved).

Fix this by checking for document.readyState and based on that if the
event has already happened, queue init() as the next thing (via rIC).

Bug: T169385
Bug: T175918
Change-Id: I774ea6880f3463c606b1d45f9cb60bfb6c5a2617