build: Remove outdated references to server/
The development version of the EventLogging server has not been
maintained in this repository since sometime in 2015.

All code related to the EventLogging server is now in
wikimedia/eventlogging.git. The MediaWiki extension in this
repository is purely an EventLogging client in JS, and PHP.

Also remove the Travis CI config as it has been failing for years
and is redundant with the Jenkins jobs already running from Gerrit.

Change-Id: I2a6eb3dfa1b7aa8e707a35718e16103709fe7579