New: add mw.eventLog.getPageviewToken()
Add a new API to mw.eventLog, getPageviewToken(), which returns a token
cached with the page for association with a given user's pageview. The
token is generated by a call to mw.user.generateRandomSessionId() which
is deemed to be adequately random for the immediate need of
getPageviewToken() in WikimediaEvent's ReadingDepth schema.

Bug: T201124
Change-Id: I9ec51d160cc4500e824d515f9b472b92912968df