Merge subscriber functionality into core module
This deprecates the subscriber module and leaves it empty, depending on
the core module.  This way, users can require only 'ext.eventLogging'
and get all the functionality they need to do mw.track or
mw.eventLog.logEvent directly.  The Schema class is also moved to the
core module, so the subscriber module can be removed once its uses are
removed.

Also changes slightly where mw.eventLog is defined.  Now that the core
module houses all the logic, it makes more sense for core.js to define
mw.eventLog and for other files to extend it.

Bug: T187207
Change-Id: Ib64f0a892c92935ab2b5125a1b6f5308a45abd38