Add eventStream (phase 1) to ext.eventLogging
Adds an EventLogging client that sends events to EventGate.

This adds a new 'submit' member to mw.eventLog:

  mw.eventLog.submit( streamName, eventData );

In order to submit events to a stream, that stream's configuration
must be loaded.  To do so, it must match an entry in the
$wgEventStreams MW config variable with its config, as well
as be listed in the $wgEventLoggingStreamNames list.

E.g.

  $wgEventStreams = [
      [
          "stream" => "my.stream",
          "schema_title" => "my/event/schema",
          "sample_rate" => 0.5 // or whatever we need to configure
      ],
      ...
  ];

  ...
  $wgEventLoggingStreamNames = [
      'my.stream',
      ...
  ];

Configuration for all streams in $wgEventLoggingStreamNames
will be loaded and registered for use.

Some tests are added as well.

This also slightly refactors the ext.eventLogging entrypoint to be
index.js instead of subscriber.js.  The index.js entrypoint is
what attaches functions to the global mw object all other
packageFiles use locally defined functions.

Bug: T244521
Change-Id: Id2b0d00f3ac335239bb906f93afb1720a0c143ce