Make BackgroundQueue more aware of page unload flow
The queue.flush() on pagehide meant that anything up to and including
`beforeunload` would be sent, but anyone doing logging within pagehide
would be execution-dependent based on the order of event registration,
and logging in unload would be skipped.

This changes it so that pagehide and unload will both trigger another
queue flush to catch anything logged after our pagehide handler, and any
logging done after pagehide will be immediately fired.

Bug: T246382
Change-Id: Ib3bba780f5cbfdffa80841847a2ba8047c0027cf