Support POSTing legacy EventCapsule style events to EventGate
Now that EventLogging can POST to EventGate, we want to be able to
transparently migrate existent instrumentation events to the
new Event Platform components.

If an EventLoggingSchema's revision entry is a string, we assume
that string is a $schema uri rather than a MediaWiki
rev_id. This schema uri will be set as the $schema field
used by EventGate to lookup the schema and validate the event.

We also prepare the encapsulated event to be compatible with
some Event Platform schema conventions.
We ensure that event.dt is used for event.meta.dt,
and that event.meta.domain is set to the value of event.webHost.

To migrate an EventLogging Schema to Event Platform:

- Copy and convert the on wiki Schema to the schemas/event/secondary
  repository, like: https://schema.wikimedia.org/#!//secondary/jsonschema/analytics/legacy/searchsatisfaction

- Add a a stream config entry in wgEventStreams that matches the stream name,
  like eventlogging_<SchemaName>.

- Set your schema's entry in wgEventLoggingSchemas to the relative $schema
  URI in the schemas/event/secondary repository.  E.g.
  {
    ...
    "SearchSatisfaction": "/analytics/legacy/searchsatisfaction/1.0.0",
    ...
  }

Bug: T238230
Bug: T249261
Change-Id: I1a20efeda7071aa36d06ddea9b0c850a7db1208a