Record event generation time in client_dt
This patch switches from using meta.dt to record
client-side timestamp of when the event was logged
to using client_dt instead. The meta.dt field is
not set by client so that EventGate can set it,
enabling it to be used as the time of receipt.

Tested in MediaWiki Vagrant & Special:JavaScriptTest

Bug: T240460
Change-Id: Ic2fcd16587660681d2ffbc12e22469893d27502d
Related-Id: I4b23cd2ba0eee92035a3eff20ac0519849ca78fd