Event Platform Client for MediaWiki PHP
Adds a new static method, EventLogging::submit, which can be used
to submit event data to the Event Platform. The client is essentially
a direct port of the Event Platform JS client implemented in the
ext.eventlogging/core module.

EventLogging::submit delegates to a new EventPlatformClient class in
which the implementation resides. The class resides under the Libs/
folder in keeping with our conceptualization of the Event Platform
Clients as a set of libraries.

Bug: T253121
Change-Id: I57dc14b822c3c85d41eca30fed9f09d70f8ca993