MEP-PHP: Provide forward compatibility of migrated event schemas
Updates EventLogging::logEvent to handle event schemas that have been
migrated to the Event Platform. As updated, we begin by loading schema
information from $wgEventLoggingSchemas. If a schema's revision entry
is a string, we assume that string is a $schema uri rather than a
MediaWiki rev_id. This schema uri will be set as the $schema field
used by EventGate to look up the schema and validate the event.

We also prepare the event to be compatible with some Event Platform
schema conventions. We add the schema name as event.$schema, add the
current timestamp as event.client_dt, and set event.meta.domain to
$wgServerName.

To migrate an EventLogging Schema to Event Platform:

- Copy and convert the on wiki Schema to the schemas/event/secondary
  repository, like: https://schema.wikimedia.org/#!//secondary/jsonschema/analytics/legacy/searchsatisfaction

- Add a a stream config entry in wgEventStreams that matches the stream name,
  like eventlogging_<SchemaName>.

- Set your schema's entry in wgEventLoggingSchemas to the relative $schema
  URI in the schemas/event/secondary repository.  E.g.
  {
    ...
    "SearchSatisfaction": "/analytics/legacy/searchsatisfaction/1.0.0",
    ...
  }

Bug: T253121
Change-Id: Iee3a49f3b98bee8356eb7baa6a403292c0a2f2ba