Provide forward compatibility of migrated event schemas (take 2)
Updates EventLogging::logEvent to handle event schemas that have been
migrated to the Event Platform. As updated, we begin by loading schema
information from $wgEventLoggingSchemas. If a schema's revision entry
is a string, we assume that string is a $schema uri rather than a
MediaWiki rev_id. This schema uri will be set as the $schema field
used by EventGate to look up the schema and validate the event.

We also prepare the event to be compatible with some Event Platform
schema conventions. We add the schema name as event.$schema, add the
current timestamp as event.client_dt, and set event.meta.domain to
$wgServerName.

NOTE: A previous version of this change broke CI for some other
projects. The cause has been identified as a missing check for the
eventlogging service base URI being configured. This has been fixed
in the current version.

To migrate an EventLogging Schema to Event Platform:

- Copy and convert the on wiki Schema to the schemas/event/secondary
  repository, like: https://schema.wikimedia.org/#!//secondary/jsonschema/analytics/legacy/searchsatisfaction

- Add a a stream config entry in wgEventStreams that matches the stream name,
  like eventlogging_<SchemaName>.

- Set your schema's entry in wgEventLoggingSchemas to the relative $schema
  URI in the schemas/event/secondary repository.  E.g.
  {
    ...
    "SearchSatisfaction": "/analytics/legacy/searchsatisfaction/1.0.0",
    ...
  }

Bug: T253121
Change-Id: Ic902d5d50797d890048a56a6152935cd56ae4f2d