Demo inclusion of client-provided data based on stream config
After checking that the stream is in sample and before sending, reads
the stream configuration and supplements the event data as requested.

For this demo implementation, the client only knows how to supplement
event data with the current MediaWiki skin, and always places it in a
top-level `test` map within the event data. The intent is to demo
this with session tick events conforming to the analytics/session_tick
schema (currently v2.0.0).

Bug: T271456
Change-Id: I8ae363edb53fbd7d8197fac9fd41ecd1041e3168