Fix timestamp format for migrated events
The event formatting code for migrated events currently uses the
date() function to generate timestamps, which causes two problems:

(1) they specify timezone in +0000 format rather than with the
preferred Z suffix;

(2) they are not guaranteed to default to UTC time.

To fix both issues, these calls are replaced with calls to
wfTimestamp.

https://www.mediawiki.org/wiki/Manual:WfTimestamp

Bug: T276235
Change-Id: Ibff3d94b797cbe8cb31dc0b667bb3dcc8fb12282