Do not set dt field for migrated legacy events (PHP)
For migrated legacy events, client_dt is provided as the client-side
timestamp, and dt should be omitted so that EventGate can set it to
the same value as meta.dt, i.e., the server intake time.

Additional change: To reduce ambiguity, treat dt as reserved for the
client library and set it unconditionally if client_dt is unset.

Bug: T277330
Change-Id: I4f065c5728b2b28520ce1615dd014367df6a4212