Unset dt in migrated legacy events (PHP)
For migrated legacy events, the `dt` field should be left unset so
that it can be set to the intake time by EventGate. This unsets it
if it is set by a caller and `client_dt` is also set, indicating that
it is a migrated legacy event.

This is primarily for future-proofing, as there are no known cases
of `dt` being set by legacy event producers.

Bug: T277330
Change-Id: I26ebc9c69852e5d26d04d3d5f700255410af7f28