Unset dt in migrated legacy events (JS)
For migrated legacy events, the `dt` field should be left unset so
that it can be set to the intake time by EventGate. This unsets it
if it is set by a caller and `client_dt` is also set, indicating that
it is a migrated legacy event.

This is primarily for future-proofing, as there are no known cases
of `dt` being set by legacy event producers.

Bug: T277330
Change-Id: I09d38e94b257ffab13a16e0720d570f240c5078d