Restore dependency-free nature of this extension
This extension has often been referred to as a dependency-free
interface that any extension may depend on without having to worry
about complicating installation for end-users that may not be
interested in logging events. E.g. configuring it and installing other
things should be optional.

To restore this, remove the hardcoded EventBus and EventStreamConfig
dependencies introduced in c4dd0ae and 21f985f, and ensure that the
extension still works correctly.

The expectation is that installing something like Echo and
VisualEditor on a plain MW install with EventLogging out-of-the-box
should results in the PHP and JS interfaces defined as no-ops.

Change-Id: Icdc9220d55c7301640013c802e427f70e0bfde56
(cherry picked from commit 2acaa7987b73dda1ee46e36f8fadfef8add6f0ae)