build: Remove eventgate from package.json
This is unrelated to the extension or the CI for the PHP/JS source
files in this repo. It is something one can install separately to
as convenient way to run a devserver. Make this separate again, like
it was for the legacy devserver.

When running the lint check against this repo (npm ci && npm test) for
checking JS files, it should not require downloading dependencies
for eventgate.

Bug: T285372
Change-Id: Ie065dc8cb1e9125c0f286d1533e1c9bc8a99b052