Allow limiting debug mode to console
Add a new user option, eventlogging-display-console, which works
similarly to eventlogging-display-web, but only logs to console
and does not show a popup notice. With most features doing large
volumes of logging, the notice approach is cumbersome to use at
best and can interfere with the functionality under test at worst,
while the console can be searched or filtered (or ignored in favor
of a handcrafted mw.track subscriber).

Change-Id: I857bc50427f4c267c21c695f60188a631086af22