MediaWikiMetricsClientIntegration: Cache ::getContextAttributes() result
In the review of I9385d615ee3, it was noted that
MetricsClient::dispatch() calls ::getContextAttributes() for each
event-stream pair. Further, if an event is dispatched to stream N times,
then ::getContextAttributes() will be called N times.

We cache the result of ::getContextAttributes() at the module level to
improve performance. Since the result of does not vary by instance, it
is safe to cache it at this level.

Change-Id: I31b9b4064e51e8394c95e23e14219e2baaded16b