MediaWikiMetricsClientIntegration: Do not read state from DOM
In the review of I9385d615ee3, it was noted that
::getContextAttributes() determines whether MobileFrontend is active by
reading state from the DOM, i.e. by checking whether <body> has the
"mw-mf" class. This was determined to be a non-trivial read from the DOM
and one that could cause a forced style recalculation in certain
situations.

We make ::getContextAttributes() determine whether MobileFrontend is
active by reading the wgMFMode config variable.

Note well that the use of the wgMFMode config variable in this way is
not recommended. ::getContextAttributes() will need to be updated at the
same time as MobileFrontend.

Change-Id: Id13a82955c823016c80222d402eb4356ddc8b5a9