Hooks: Filter stream config settings
Add an allowlist that is used to filter stream config settings that
should not be sent to the client as part of the ext.eventLogging RL
module. Because the allowlist is unlikely to change frequently, it is
sufficient to make it a private class constant rather than a config
variable.

The initial value of the allowlist is: [ 'sample', 'producers' ]

This filter supersedes the one in
MediaWiki\Extension\EventStreamConfigs\StreamConfig. Both
are responsible for minimizing the size of the ext.eventLogging RL
module but that responsibility belongs to this codebase only.

Bug: T286344
Change-Id: Ibc31826a20ba303368f8ae8e92fbb3b4d58b447a