Extract UserBucketService service class
Extract the Libs\UserBucketProvider\UserBucketService service class from
UserBucketProvider. As well as providing a method equivalent to
UserBucketProvider::getUserEditCountBucket(), the service class also has
a method that buckets an integer edit count, which is useful in the case
where you have already accessed the edit count of a user elsewhere.

Bug: T281762
Change-Id: Ibef9e7975bdcdbc2bc9a7e43049366bf6b56c294