Extract EventSubmitter interface
The primary responsibility of the EventLogging and EventLoggingHelper
classes is to submit events to some event ingestion service. Currently,
all event sources have to invoke EventLogging::submit() directly, which
breaks DI.

Create the EventSubmitter interface, which defines the API for
submitting events to one or more arbitrary ingestion service. Extract
the current functionality of the EventLogging and EventLoggingHelper
classes into the EventBusEventSubmitter class.

Bug: T281762
Change-Id: If91e9b321edc3d5f414aafaf8ec13cf0735a7dd4