tests: Make test independent of developer's content and user lang
Improve test by expecting concrete values instead of duplicating
the source, which merely tests that we have managed to write the
same code twice, not that it produces the expected result.

Without this, we can't actually test whether the code is calling the
correct methods or that those methods are valid to call in the way
that it does. E.g. if both return null we little further than having
no test coverage.

While at it, extend MediaWikiLangTestCase so that the test runs with
'en' as the content and user language, instead of assuming
LocalSettings.php is configured as English.

Change-Id: I8c2b8c20b805db01d1c093f09b1fe4a298b07bb8