ContextAttributesFactory: Use Title->getPageLanguage() to avoid global state
This method takes $context as parameter, but getPageViewLanguage()
bypasses that in a non-trivial way to inspect the global state
directly.

Calling getPageLanguage returns the same result as before for 99%
of pages and page views, including the case where it inspects global
state to return the UI language on Special pages.

What it changes, is, a subset of page views on wikis like
zh.wikipedia.org where there exist language variants. Previously, if
you viewed zhwiki with a language preference that differs from the
default, but is still within `zh-*`, then getPageViewLanguage() would
return your personal language preference instead of the page content
language.

For any other language preference, and on any other wikis, the two
methods return the same value.

Bug: T350806
Change-Id: I5149024122be0eb8d8fc6c02920a401fd9dd6a67