tests: Update EventLoggingTest::testDispatch()
... in preparation for updating wikimedia/metrics-platform to 3.1.0

EventLoggingTest::testDispatch() and related tests are integration tests
that exercise the integration of EventLogging and the
wikimedia/metrics-platform library. These tests do catch regressions but
at the cost of increased friction when updating the library. For
instance, the latest version of the library will always include
information about the agent and these tests must be updated to reflect
that prior to updating the mediawiki/vendor repository.

Bug: T351294
Change-Id: I46e513115f942f5b1f7e32881d27ddae3fe6a310