jsonschemacontent: Remove broken objectRow "$ref" handling
This existed to handle "$ref" values containing "/123",
which would render a linked permalink to that number
as a local revision ID.

This was added in 2014 (I2b5aae432d8f8259d0bc170f69f36),
broken in 2015 (I8910872fea84eeb) and is unused today.

Removing this only removes the custom pretty rendering
for it, it does not affect whether this can be stored
and retreived from Json pages.

To reproduce:

1. $val could be an instance of stdClass

   For example, in [0], $val is an instance of stdClass with
   "description", "type", and "title" properties

2. $val could be an URI fragment

   For example, in [0] and [1] $val is "#schemanode"

[0]: https://meta.wikimedia.org/w/index.php?title=Schema:Schema&oldid=4867919&action=raw
[1]: https://meta.wikimedia.org/w/index.php?title=Schema:Schema&oldid=27016577&action=raw

Bug: T368543
Change-Id: I68e5a0d2108acea1aa7e969953bd7c7e1beefe94