ext.eventLogging: Extract ext.eventLogging.metricsPlatform module
What?

We extract the ext.eventLogging.metricsClient module, which provides
developers the ability to create a specialized MetricsClient instance.

For now, all MetricsClient instances use the same stream configs. The
stream configs are now delivered as part of the config for the
ext.eventLogging.metricsClient module, which leaves the ext.eventLogging
module agnostic to stream configs.

Why?

Experiment-related analytics events need to be sent to a dedicated event
intake service. The intake service is responsible for setting the
subject ID on events related to everyone experiments. It is also
responsible for rejecting events that aren't related to experiments (see
T391959 for more detail). That is, it would drop all other analytics
events.

Bug: T391988
Change-Id: I3d78ec8749cceb851d21930facfd3ef7376dab9b